# api/openapi/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.SearchVideoData:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
      videos:
        items:
          $ref: '#/definitions/dto.SearchVideoInfo'
        type: array
    type: object
  dto.SearchVideoInfo:
    properties:
      author_id:
        type: integer
      author_name:
        type: string
      comment_count:
        type: integer
      cover_url:
        type: string
      description:
        type: string
      favorite_count:
        type: integer
      highlight:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      id:
        type: integer
      play_url:
        type: string
      publish_time:
        type: integer
      title:
        type: string
      view_count:
        type: integer
    type: object
...
  dto.UserInfo:
    properties:
      avatar:
        type: string
      background_image:
        type: string
      follow_count:
        type: integer
      follower_count:
        type: integer
      id:
        type: integer
      user_name:
        type: string
      user_role:
        type: string
    type: object
  dto.UserUpdateRequest:
    properties:
      avatar:
        maxLength: 500
        type: string
      background_image:
        maxLength: 500
        type: string
      username:
        maxLength: 255
        minLength: 1
        type: string
    type: object
...
  response.ErrorInfo:
    properties:
      code:
        type: integer
      message:
        type: string
      type:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorInfo'
    type: object
  response.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: 127.0.0.1:8000
info:
  contact:
    email: support@vida.com
    name: API Support
  description: 视频分享平台 API 服务
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Vida-Go API
  version: "1.0"
paths:
...
  /search/sync:
    post:
      description: 将数据库中的视频同步到 Elasticsearch
      produces:
      - application/json
      responses:
        "200":
          description: 同步成功
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 同步失败
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: 同步视频到ES
      tags:
      - 搜索
  /search/videos:
    get:
      description: 根据关键词搜索视频，支持多种筛选条件
      parameters:
      - description: 搜索关键词
        in: query
        name: q
        type: string
      - description: 作者ID
        in: query
        name: author_id
        type: integer
      - description: 视频ID
        in: query
        name: video_id
        type: integer
      - description: 视频状态
        in: query
        name: status
        type: string
      - default: relevance
        description: '排序方式: relevance, latest, hot'
        in: query
        name: sort
        type: string
      - description: 开始时间戳
        in: query
        name: start_time
        type: integer
      - description: 结束时间戳
        in: query
        name: end_time
        type: integer
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 搜索成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.SearchVideoData'
              type: object
        "400":
          description: 请求参数无效
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: 搜索视频
      tags:
      - 搜索
  /users:
    get:
      description: 分页获取用户列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 用户名筛选
        in: query
        name: username
        type: string
      - description: 角色筛选
        in: query
        name: user_role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取用户列表（管理员）
      tags:
      - 用户
  /users/{id}:
    delete:
      description: 软删除指定用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除用户（管理员）
      tags:
      - 用户
    get:
      description: 根据用户ID获取用户信息（需要权限）
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserInfo'
              type: object
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取指定用户信息
      tags:
      - 用户
    put:
      consumes:
      - application/json
      description: 更新指定用户的信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserInfo'
              type: object
        "400":
          description: 请求参数无效
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新用户信息
      tags:
      - 用户
  /users/{id}/restore:
    post:
      description: 恢复已删除的用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 恢复成功
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 恢复用户（管理员）
      tags:
      - 用户
  /users/{id}/set-admin:
    post:
      description: 将指定用户设置为管理员
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 设置成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserInfo'
              type: object
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 设置管理员（管理员）
      tags:
      - 用户
  /users/me:
    get:
      description: 获取当前登录用户的信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserInfo'
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取当前用户信息
      tags:
      - 用户
...
securityDefinitions:
  BearerAuth:
    description: '输入格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"